initial changes for NFIRAOS FDR
Bringing more in line with recent TCS - IRIS ICD changes.

Pointingstage and trackID now included in device-specific demands
so that we don't usually need 3 separate event streams (current,
target, pointingStatus). Exception is that we have left them
separate in this commit for the pk assembly as these are generic
streams that may be used in different ways by different subsystems
in TMT (i.e., some devices may want to know where we are now, whereas
others want to prime for the target location).